refactor: extract OpenAI API and tool helpers
